pyproject.toml
[project]
name = "rag-playground"
version = "1.0.0"
description = "Daily podcast combining differents tecnhiques based on LLMs models"
authors = []
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-groq (>=0.2.4,<0.3.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "elevenlabs (>=1.50.6,<2.0.0)",
    "crawl4ai (>=0.4.248,<0.5.0)",
    "litellm (>=1.60.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false